django_backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
python-decouple==3.8
bcrypt==4.0.1
pillow==11.3.0
PyJWT==2.7.0